cosmetic changes
Use the "bakery" identifier for the macaroonbakery package everywhere.

Use all-args-on-separate-lines style when we want to split
a function onto multiple lines, as it's less sensitive to
automated changes that change the length of the initial
function call.

Split multiple import-from statements across multiple lines
and sort the imported identifiers.