Improve mock setup for 407-then-unknown test
`test_407_then_unknown_interaction_methods` causes the client to fetch
the possible methods supported by the discharger (because it's told that
it only supports a non-window method).  This is currently unmocked,
which causes the client to actually contact `http://example.com/visit`.
This fails in Launchpad builds because they run with a restrictive
network setup that doesn't even expose DNS lookups for non-permitted
hosts.

There isn't really a good way to simulate this without setting up a
similar stunt DNS server (though perhaps installing an
`httmock.all_requests` fallback mock that raises an exception would be a
good idea?), but this seems to be the only failure at the moment.