bakery/checker: fix verification exception behavior
The current logic is backwards, because it immediately raises an
exception if it finds that any macaroon is bad, where it should
raise an exception only if there's a problem that *isn't* because
a macaroon is bad.

The AuthInitError exception should be (and now is) reserved for
problems that are not because of the input macaroons themselves.

The allow methods should never raise a VerificationError.